requirements.txt
streamlit>=1.20.0
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
requests>=2.31.0